Ready to play with
This has a dummy circuit built on it, with a large voltage regulator at
the bottom, needing to be soldered to a copper area with the "SOLID"
property, that will act as a heat sink.  The question is, how do you
join it with the copper pour at the top, that has the area property of
"THERMAL RELIEF"?  If they don't physically connect, then the GROUND
connection of the voltage regulator has no path.