Added GND plane "Priority" property
I now have to GND planes, both on the front side copper.  The upper
plane, is set to priority "2", and the bottom is set to priority "1".
The upper is set to thermal relief (check the ground pad on the pin
header, upper right), and the bottom is set to solid, to ensure full
contact with the thermal pad on the DPAK3 package.  Even though this
would result in a fully conductive surface in reality, you have to add a
teeny trace to connect the two planes for KiCad to consider them
attached.

Try deleting the teeny trace; you'll see that KiCad returns the ratsnest
line between the GND pad of the DPAK3 at the bottom and the top
component.